requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.12.14
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
Django==4.2
django-extensions==3.2.3
exceptiongroup==1.2.2
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
matplotlib-inline==0.1.7
openai==1.60.0
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
shell==1.0.1
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
